Fixed issue with having non AnyCPU platforms
This didn't take into account the solution configurations when calculating the output path
This now uses the solution configuration to pass through Configuration|Platform to get the output path.
Also uses the PackageOutputPath, rather than the OutputPath and then modify. The PackageOutputPath is the OutputPath without the TFM appended.
Fixes issue #1